Restructure the current dell driver: * A super class (_base_session) that has methods common to dell switches. * The driver for powerconnect inherits from this class.
